Evaluator package improvements
